Added DataSourceInfo capabilities: isRandomAccessEfficient(), canHandleStrides() Strides vector has changed from vector<int> to vector<long> Nifti reader can handle strides, which fixes #10438
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/soma/soma-io/trunk@62511 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/libsoma-io/image/imagereader.h
@@ -78,7 +78,7 @@ namespace soma
       virtual void read( T * dest, DataSourceInfo & dsi,
                          std::vector<int> & pos,
                          std::vector<int> & size,
-                         std::vector<int> & stride,
+                         std::vector<long> & stride,
                          carto::Object options = carto::none() );
 
       /// Abstract : set specialized ImageReader's parameters.
