+++ src/somadicom/image/dicomimagereader_d.h
@@ -103,7 +103,7 @@ namespace soma {
   void DicomImageReader<T>::read( T * dest, DataSourceInfo & dsi,
                                   std::vector<int> & /* pos */,
                                   std::vector<int> & /* size */,
-                                  std::vector<int> & /* stride */,
+                                  std::vector<long> & /* stride */,
                                   carto::Object      /* options */ )
   {
     if ( dest )
