+++ src/somagis/image/gisimagereader_d.h
@@ -216,7 +216,7 @@ namespace soma {
   void GisImageReader<T>::read( T * dest, DataSourceInfo & dsi,
                                 std::vector<int> & pos,
                                 std::vector<int> & size,
-                                std::vector<int> & /* stride */,
+                                std::vector<long> & /* stride */,
                                 carto::Object      /* options */ )
   {
     if( _sizes.empty() )
